feat: add authentication guards to user pages
- Protect /dashboard/billing page with requireAuth()
- Protect /favorites page with requireAuth()
- /submit page remains public (no auth required)
- Convert billing page from client to server component
