+++ app/[locale]/dashboard/billing/page.tsx
@@ -1,10 +1,12 @@
-"use client";
-
 import { useTranslations } from "next-intl";
 import { PageContainer } from "@/components/ui/container";
 import { BillingSection } from "@/components/dashboard";
+import { requireAuth } from "@/lib/auth/guards";
 
-export default function DashboardBillingPage() {
+export default async function DashboardBillingPage() {
+  // Require authentication
+  await requireAuth();
+
   const t = useTranslations();
   return (
     <div className="min-h-screen bg-gradient-to-br from-gray-50 via-white to-gray-100 dark:from-gray-900 dark:via-gray-800 dark:to-gray-900">
